Update README.md
Created an optimization model for a stock portfolio that uses PyPortfoliopt library and Efficient Frontier to allocate investment amounts in such a way that the portfolio has decent returns with less volatility. The efficient frontier model works on the Sharpe Ratio of the portfolio. For an investment amount, the model gives the weights (i.e. the number of shares per stock) that need to be bought in order to optimize the portfolio.